requirements.txt
llama-index-readers-llama-parse==0.1.4
python-dotenv==1.0.1